requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-filter>=23.2
djangorestframework-simplejwt>=5.3.1

# Swagger
drf-spectacular>=0.26.0

# Database
psycopg2-binary>=2.9.6

python-dotenv>=1.0.0

# Pytest
pytest>=7.0
pytest-django
